requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.28.0
datasets>=2.12.0
numpy>=1.22.0
pandas>=1.5.0
Pillow>=9.4.0
albumentations>=2.0.6

editdistance>=0.6.0
torchmetrics>=0.11.0

matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tqdm>=4.65.0

pyyaml>=6.0
unicodedata2>=15.0.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
jiwer
flask